+++ src/webserver/database/sqlpage_functions/mod.rs
@@ -1,12 +1,9 @@
-mod function_definition_macro;
 mod function_traits;
 pub(super) mod functions;
 mod http_fetch_request;
 mod url_parameters;
 
 use sqlparser::ast::FunctionArg;
-
-use crate::webserver::http_request_info::{ExecutionContext, RequestInfo};
 
 use super::sql::ParamExtractContext;
 use super::syntax_tree::SqlPageFunctionCall;